Fixes to the run-tests.sh script.
The run-tests.sh script now copies all files from the student's code
directory into a temporary directory so that it can fix permissions on
those files. This is useful in circumstances where the student may have
set different permissions on their files (e.g. umask 0077) that would
cause us to be unable to read the files during testing.

In addition, the script now checks for the presence of an XFUZZ_COMMAND
environment variable. If it finds that variable, it runs it in place of
the default testing script.